feat: Adds option to generate and a conf id string for client tokens.
The same way we generate a string for the e2ee key.
